M1590: move summary site materialized view to a model
